requirements.txt
fastapi
python-dotenv
PyJWT
pymongo
fastapi-utils
bcrypt
uvicorn
dnspython
requests
pendulum